requirements.txt
requests
pandas
pytubefix
youtube-transcript-api
python-dotenv
